change styles

+++ client/js/routes/profile/controller.js
@@ -1,9 +1,8 @@
 /* global angular */
 
 angular.module('movTv')
-  .controller('ProfileController', function ($scope, sweetAlert, $rootScope, $routeParams, FavoritesService, CommentsService, MediaService) {
+  .controller('ProfileController', function ($scope, $window, sweetAlert, $rootScope, $routeParams, FavoritesService, CommentsService, MediaService) {
     $scope.user = $routeParams.user
-    console.log($scope.users)
     loadFavorites()
     loadComments()
     $scope.editFavId = ''
@@ -14,14 +13,13 @@ angular.module('movTv')
       FavoritesService.getFavoritesByUser($scope.user)
       .then(function (response) {
         $scope.favorites = response.data
-        console.log($scope.favorites)
         $scope.favorites.forEach(favorite => MediaService.searchDetail(favorite.imdbID)
         .then(function (response) {
           response.data.stars = favorite.stars
+          console.log(response.data.stars)
           response.data.favId = favorite._id
           $scope.FullFavorites.push(response.data)
         }))
-        console.log($scope.FullFavorites)
       })
     }
 
@@ -29,15 +27,15 @@ angular.module('movTv')
       if ($scope.user !== $rootScope.loggedUser) {
         sweetAlert.swal({
           type: 'error',
-          text: `Sorry you can't edit another user's profile!`
+          text: `Sorry you can't edit another user's collection!`
         })
       } else {
         sweetAlert.swal({
-          title: 'Delete FAvorite',
+          title: 'Delete',
           text: 'Are you sure you want to delete this from your collection?',
           type: 'error',
           showCancelButton: true,
-          confirmButtonColor: '#3085d6',
+          confirmButtonColor: '#0CE3AC',
           cancelButtonColor: '#d33',
           confirmButtonText: 'Confirm',
           cancelButtonText: 'Cancel'
@@ -46,7 +44,7 @@ angular.module('movTv')
           FavoritesService.removeFavorite($scope.user, imdbID)
             .then(loadFavorites())
         }, function (dismiss) {
-          console.log('cancelled')
+          // do nothing
         })
       }
     }
@@ -59,12 +57,11 @@ angular.module('movTv')
       if ($scope.user !== $rootScope.loggedUser) {
         sweetAlert.swal({
           type: 'error',
-          text: `Sorry you can't edit another user's profile!`
+          text: `Sorry you can't edit another user's collection!`
         })
       } else {
-        console.log($scope.editFavId)
         FavoritesService.editStarsFav($scope.editFavId, newStars)
-        .then(loadFavorites())
+          .then($window.location.reload())
       }
     }
 
@@ -76,13 +73,19 @@ angular.module('movTv')
           .then(function (response) {
             comment.movieTitle = response.data.Title
           }))
-        console.log($scope.Comments)
       })
     }
 
     function removeComment (commentId) {
-      CommentsService.removeComment(commentId)
+      if ($scope.user !== $rootScope.loggedUser) {
+        sweetAlert.swal({
+          type: 'error',
+          text: `Sorry you can't edit another user's collection!`
+        })
+      } else {
+        CommentsService.removeComment(commentId)
         .then(loadComments())
+      }
     }
 
     $scope.removeFavorite = removeFavorite
